Move lock middleware to file and add to projectTypes API
